feat(auth): refactor ForgetPassword UI hard coded Strings
