+++ lib/core/l10n/app_en.arb
@@ -1,96 +1,80 @@
 {
   "@@locale": "en",
-
   "appName": "Exam App",
   "signup": "Sign Up",
-
   "userNameLabel": "Username",
   "userNameHint": "Enter username",
   "emptyUserNameError": "Username is not valid",
-
   "firstNameLabel": "First Name",
   "firstNameHint": "Enter first name",
   "emptyFirstNameError": "First name is not valid",
-
   "lastNameLabel": "Last Name",
   "lastNameHint": "Enter last name",
   "emptyLastNameError": "Last name is not valid",
-
   "emailLabel": "Email",
   "emailHint": "Enter email",
   "emptyEmailError": "Email is not valid",
-
   "passwordLabel": "Password",
   "passwordHint": "Enter password",
   "emptyPasswordError": "Password is not valid",
-
   "confirmPasswordLabel": "Confirm Password",
   "confirmPasswordHint": "Confirm password",
   "emptyConfirmPasswordError": "Passwords do not match",
-
   "phoneLabel": "Phone",
   "phoneHint": "Enter phone number",
   "emptyPhoneError": "Phone number is not valid",
-
   "signupButtonText": "Sign Up",
   "alreadyHaveAccount": "Already have an account?",
   "login": "Login",
-  "wait":"Please wait...",
+  "wait": "Please wait...",
   "welcome": "Welcome",
-  
   "tooShort": "The length should be at least 3 characters",
   "invalidName": "Enter a valid name",
   "invalidEmail": "Enter a valid email",
   "invalidPassword": "Password must be at least 8 characters,\ninclude upper, lower, digit & special char",
   "confirmPasswordMismatch": "The Confirm password is not matched with your password",
   "invalidPhoneNumber": "Enter valid Egyptian phone number",
-
-  
-    "email": "Email",
-    "enterYourEmail": "Enter your email",
-    "emailError": "This Email is not valid",
-    "password": "Password",
-    "pleaseEnterYourEmail" :"Please enter your email",
-    "pleaseEnterValidEmail" :"Please enter valid email",
-    "enterYourPassword": "Enter your password",
-    "rememberMe": "Remember me",
-    "forgetPassword": "Forget password?",
-    "dontHaveAccount": "Don't have an account?",
-    "signUp": "Sign up",
-  
-    "userName": "User name",
-    "enterUserName": "Enter your user name",
-    "passwordMinLengthError": "Password must be at least 6 characters",
-    "userNameError": "This user name is not valid",
-    "firstName": "First name",
-    "enterFirstName": "Enter first name",
-    "lastName": "Last name",
-    "enterLastName": "Enter last name",
-    "confirmPassword": "Confirm password",
-    "enterConfirmPassword": "Confirm password",
-    "passwordNotMatched": "Password not matched",
-    "phoneNumber": "Phone number",
-    "enterPhoneNumber": "Enter phone number",
-    "forgetPasswordTitle": "Forget password",
-    "forgetPasswordSubtitle": "Please enter your email associated to your account",
-    "continue_button": "Continue",
-    "emailVerificationTitle": "Email verification",
-    "emailVerificationSubtitle": "Please enter your code that send to your email address",
-    "invalidCode": "Invalid code",
-    "didntReceiveCode": "Didn't receive code?",
-    "resend": "Resend",
-    "resetPasswordTitle": "Reset password",
-    "resetPasswordSubtitle": "Password must not be empty and must contain 6 characters with upper case letter and one number at least",
-    "newPassword": "New password",
-    "emailVerification":"Email verification",
-    "enterCodeToSendToEmailAddress":"Please enter your code that send to your \n email address ",
-    "donotReciveCode":"Didn't receive code? ",
-    "resetPassword":"Reset Password",
-    "passwordruleAtResetPassPage":"Password must not be empty and must contain 6 characters with upper case letter and one  number at least",
-    "forgetPass":"Forget Password",
-    "pleaseEnteryouEmailAtForgotPassPage": "Please enter your email associated to \n your account"
-
-
-    
-  
+  "email": "Email",
+  "enterYourEmail": "Enter your email",
+  "emailError": "This Email is not valid",
+  "password": "Password",
+  "pleaseEnterYourEmail": "Please enter your email",
+  "pleaseEnterValidEmail": "Please enter valid email",
+  "enterYourPassword": "Enter your password",
+  "rememberMe": "Remember me",
+  "forgetPassword": "Forget password?",
+  "dontHaveAccount": "Don't have an account?",
+  "signUp": "Sign up",
+  "userName": "User name",
+  "enterUserName": "Enter your user name",
+  "passwordMinLengthError": "Password must be at least 6 characters",
+  "userNameError": "This user name is not valid",
+  "firstName": "First name",
+  "enterFirstName": "Enter first name",
+  "lastName": "Last name",
+  "enterLastName": "Enter last name",
+  "confirmPassword": "Confirm password",
+  "enterConfirmPassword": "Confirm password",
+  "passwordNotMatched": "Password not matched",
+  "phoneNumber": "Phone number",
+  "enterPhoneNumber": "Enter phone number",
+  "forgetPasswordTitle": "Forget password",
+  "forgetPasswordSubtitle": "Please enter your email associated to your account",
+  "continue_button": "Continue",
+  "emailVerificationTitle": "Email verification",
+  "emailVerificationSubtitle": "Please enter your code that send to your email address",
+  "invalidCode": "Invalid code",
+  "didntReceiveCode": "Didn't receive code?",
+  "resend": "Resend",
+  "resetPasswordTitle": "Reset password",
+  "resetPasswordSubtitle": "Password must not be empty and must contain 6 characters with upper case letter and one number at least",
+  "newPassword": "New password",
+  "emailVerification": "Email verification",
+  "enterCodeToSendToEmailAddress": "Please enter your code that send to your \n email address ",
+  "donotReciveCode": "Didn't receive code? ",
+  "resetPassword": "Reset Password",
+  "passwordruleAtResetPassPage": "Password must not be empty and must contain 6 characters with upper case letter and one  number at least",
+  "forgetPass": "Forget Password",
+  "pleaseEnteryouEmailAtForgotPassPage": "Please enter your email associated to \n your account",
+  "passwordResetSuccessfully": "Password reset successfully"
 }
